backend/requirements.txt
django-cors-headers==4.7.0
djangorestframework==3.15.0
requests==2.31.0
cryptography==41.0.0
djangorestframework-simplejwt==5.4.0
django-allauth==0.55.0
gunicorn
